+++ kubernetes/mongo-deployment.yml
@@ -1,13 +1,13 @@
 ---
-apiVersion: apps/v1beta2
+apiVersion: apps/v1beta1
 kind: Deployment
 metadata:
   name: mongo
   labels:
     app: reddit
     component: mongo
-    comment-db: "true"
     post-db: "true"
+    comment-db: "true"
 spec:
   replicas: 1
   selector:
@@ -20,8 +20,8 @@ spec:
       labels:
         app: reddit
         component: mongo
-        comment-db: "true"
         post-db: "true"
+        comment-db: "true"
     spec:
       containers:
       - image: mongo:3.2
@@ -31,4 +31,5 @@ spec:
           mountPath: /data/db
       volumes:
       - name: mongo-persistent-storage
-        emptyDir: {}
+        persistentVolumeClaim:
+          claimName: mongo-pvc-dynamic
