+++ kubernetes/tls.crt
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE-----
+MIIDCDCCAfCgAwIBAgIJAL73FAdpHJtDMA0GCSqGSIb3DQEBCwUAMBkxFzAVBgNV
+BAMMDjM1LjE4Ni4yMTAuMTMxMB4XDTE4MDQxMTEyMjU1NVoXDTE5MDQxMTEyMjU1
+NVowGTEXMBUGA1UEAwwOMzUuMTg2LjIxMC4xMzEwggEiMA0GCSqGSIb3DQEBAQUA
+A4IBDwAwggEKAoIBAQDOx6ItN7YBAT5cFUT8MQ5/gi2sh6YTI0iJc8UgLNsXu6Lf
+horpiTc57F+egamuznx+seDHnLzUP6FOUHpu6b3fkq/F2v70uD+dYaEQNF7dxH51
+NxoUkiFvknPTpHwTQ4Ft/4sTRQxxlCJXEVkM5O3YGu41WiWBlUVnOGcSpzXDxl0D
+KReLNDE1UXJD0rSEljNcbXoM3Q+/PNFmpMQFfUJWdLt3suKHSSq7mDQxLMCg9Dlh
+glYE7k7dMypepr+dul/HvNMmEmkwqE0v/HF8Dlso/rCi1vn9GYhA16euhOlXcmkX
+yGLrx7ln9+Hl/nfZ2PBHX+/4M5ktPS2ZtY8p1ZZLAgMBAAGjUzBRMB0GA1UdDgQW
+BBSdQe1DLw7tgcyv6pZ6VJN2zAOmrTAfBgNVHSMEGDAWgBSdQe1DLw7tgcyv6pZ6
+VJN2zAOmrTAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQAAJHFh
+cytOVW9w/r5e7tQKfAc3aZrzdtwhwU44XJftor+Pt0CenRcmdQvh/F/1QsLjkC4b
+9AkeF/L27bl/bQ+hZGi6IPlnxNzlAkjjQLuqQkSEZdduk343x5azaBrAbIGl4G9Q
+RCIQKr6sW0WW2UgRJNGmZTDRva11Gg1PX6Ts7Kix/THwm5HiOsbdsfiEgBba8RT3
+o7sSUR88gRVHJiiPGK9BSogHCVTFbPRqjaqYk5KaOVA2IzuuwdGAkMObqeP6EY+U
+WdkcszfVmpWWtPjyGch953+u6btGaswPrSuqu4h/5tp3WI0rYO78wcqsuy8q0SU4
+4FFFJUpLUdyo9ab9
+-----END CERTIFICATE-----
