+++ kubernetes/tls.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDOx6ItN7YBAT5c
+FUT8MQ5/gi2sh6YTI0iJc8UgLNsXu6LfhorpiTc57F+egamuznx+seDHnLzUP6FO
+UHpu6b3fkq/F2v70uD+dYaEQNF7dxH51NxoUkiFvknPTpHwTQ4Ft/4sTRQxxlCJX
+EVkM5O3YGu41WiWBlUVnOGcSpzXDxl0DKReLNDE1UXJD0rSEljNcbXoM3Q+/PNFm
+pMQFfUJWdLt3suKHSSq7mDQxLMCg9DlhglYE7k7dMypepr+dul/HvNMmEmkwqE0v
+/HF8Dlso/rCi1vn9GYhA16euhOlXcmkXyGLrx7ln9+Hl/nfZ2PBHX+/4M5ktPS2Z
+tY8p1ZZLAgMBAAECggEAfdpjisHttI7dNY58fkIdYN4KjofBPMSs0Xv07rBeaQN7
+DjEFOz6Yrlh/mimw7QQgIbyB4vwysbxIoTG6UudQ+VJ+SR4RL0E09sUFEqfNVYzS
+quTuqraGQzIP7TCfk9nyXNC8BO5xtG9Et1mHqaFsYauCcFZ9nXEM3NYnZeL/hjeN
+L1JdeyhVLRMuNRZpLn5JckC6pv0aQWyqNUyiLpzAPLJYjjZfcmMw27ijt2SHHvOm
+TMbwIWbIvGihm+wH73wU00B0JAavkHHrzWHTTtH122uA9otSDkYrYQFvg0flcNmU
+1UbaQhRwxO3V78lrxEoJ/Uet4029zSHizryWub+pAQKBgQDob3iJVQzYfS6auffQ
+jtlG2yJWtnqsyyx4Yq2LaptwITT26xzSDwgQnFKvIbL745wx+9HpZy32F8bo1zjg
+UIp5AzqxyjpbOFzAPlOr3UZcIyw663aols8C35ybVR0kpd8FzF7HKd9w79oFzRky
+lBPwv0debqrn41xz/wPdN9GCqQKBgQDjvk7RSErFiInD30Lsjd96hLg2SPw0V7qh
+tKWjosvALVNY7daRLwVrdSj56yyOmdj4gTol6eav1GgO+j+9pha3H/gSIdvUuijd
+w3cRF2e/edmr0cpxEYVm1ed1D2w7CUbzrjnCT4WVA5NDuWZ64979ObwT+4uPiYvN
+hYtbi2Pd0wKBgQCNeoRg+07lW1UJwxaPcNPaKDo/k7UaBlVcNZwy10Ijv2KwxS12
+wsWC7tP5W8bJQaLqXcBIlJBKUfC1+4LYLs4KmKeaXmCk86qpLYbafsZ8AMDEYFyK
+g7kfo6J8fhNOIlSc/+fUzKz/3k7T5yJmQfHdQWCcnzoSCbdVt6zPc1FoeQKBgG1N
+w0pSpgBUXbpyhw49sUp+NGNpeBIvINXHe9AjrCKaHmyKCfc4WCOQ4Y/BPWtbj3UF
+oMDdl+g/y4mUtc7xzekqIMejFHzus/pre9PqJdMSSlRapgzKWfpXI3VdZ5aWB0Wh
+EvudawFv69gPPY4xm65xbUEx83N6kmybcZ0IM3AdAoGBAKAyUrLJsMyG6w4MyEjz
+wNFxVbFKft6QQGviO/tEO+1lFbYX6gceWMIcL6uW9VIzWSAtqDNEGi89g+NxYMOL
+KHMwhcoCI/Z7W3JznGY3zAuVvUCclXs1dQXDWRE6yFaHE4KQu+e3Rhy3DYAS7TOx
+NUmu0AGTLMMqfzwLocjWTphf
+-----END PRIVATE KEY-----
