+++ src/comment/Gemfile
@@ -1,9 +1,10 @@
 source 'https://rubygems.org'
 
-gem 'sinatra'
+gem 'sinatra', '~> 2.0.1'
 gem 'bson_ext'
 gem 'mongo'
 gem 'puma'
 gem 'prometheus-client'
 gem 'rack'
 gem 'rufus-scheduler'
+gem 'json'
