+++ src/post-py/helpers.py
@@ -1,16 +1,9 @@
-import structlog
 from pymongo import MongoClient
 from pymongo.errors import ConnectionFailure
 from json import dumps
-from flask import request
 
-
-log = structlog.get_logger()
-
-
-def http_healthcheck_handler(mongo_host, mongo_port, version):
-    postdb = MongoClient(mongo_host, int(mongo_port),
-                         serverSelectionTimeoutMS=2000)
+def health(mongo_host, mongo_port):
+    postdb = MongoClient(mongo_host, int(mongo_port), serverSelectionTimeoutMS=2000)
     try:
         postdb.admin.command('ismaster')
     except ConnectionFailure:
@@ -19,21 +12,15 @@ def http_healthcheck_handler(mongo_host, mongo_port, version):
         postdb_status = 1
 
     status = postdb_status
+
+    with open('VERSION') as f:
+        version = f.read()
+
     healthcheck = {
         'status': status,
         'dependent_services': {
             'postdb': postdb_status
         },
-        'version': version
+        'version': version.rstrip('\n')
     }
     return dumps(healthcheck)
-
-
-def log_event(event_type, name, message, params={}):
-    request_id = request.headers['Request-Id']
-    if event_type == 'info':
-        log.info(name, service='post', request_id=request_id,
-                 message=message, params=params)
-    elif event_type == 'error':
-        log.error(name, service='post', request_id=request_id,
-                  message=message, params=params)
