+++ src/with_zipkin/comment/Dockerfile
@@ -1,0 +1,17 @@
+FROM ruby:2.2
+
+ENV APP_HOME /app
+ENV COMMENT_DATABASE_HOST comment_db
+ENV COMMENT_DATABASE comments
+
+RUN apt-get update -qq && apt-get install -y build-essential
+RUN mkdir $APP_HOME
+
+# ADD Gemfile* $APP_HOME/
+COPY . $APP_HOME
+
+WORKDIR $APP_HOME
+
+RUN bundle install
+
+CMD ["puma"]
