+++ src/with_zipkin/comment/Gemfile
@@ -1,13 +1,9 @@
 source 'https://rubygems.org'
 
-gem 'sinatra', '~> 2.0.1'
-gem 'sinatra-contrib'
-gem 'haml'
+gem 'sinatra'
 gem 'bson_ext'
-gem 'rest-client'
+gem 'mongo'
 gem 'puma'
 gem 'prometheus-client'
 gem 'rack'
 gem 'rufus-scheduler'
-gem 'tzinfo-data'
-gem 'json'
