+++ src/with_zipkin/comment/VERSION
@@ -1,0 +1,1 @@
+0.0.3
