+++ src/with_zipkin/post-py/VERSION
@@ -1,0 +1,1 @@
+0.0.2
