+++ src/with_zipkin/post-py/docker_build.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+echo `git show --format="%h" HEAD | head -1` > build_info.txt
+echo `git rev-parse --abbrev-ref HEAD` >> build_info.txt
+
+docker build -t $USER_NAME/post .
