+++ src/with_zipkin/ui/Gemfile
@@ -1,13 +1,13 @@
 source 'https://rubygems.org'
 
-gem 'sinatra', '~> 2.0.1'
+gem 'sinatra'
 gem 'sinatra-contrib'
 gem 'haml'
 gem 'bson_ext'
-gem 'rest-client'
+gem 'faraday'
 gem 'puma'
 gem 'prometheus-client'
 gem 'rack'
 gem 'rufus-scheduler'
 gem 'tzinfo-data'
-gem 'json'
+gem 'zipkin-tracer'
